+++ .gitignore
@@ -1,5 +1,8 @@
 .gradle
-/build/
+/build
+/gradle
+.idea/
+out/
 
 # Ignore Gradle GUI config
 gradle-app.setting
